Make simulated input data more realistic
The floating point case now mimics 4 bit data that is cast on the CPU
to floating point.  This results in input values in the set:

{-7.0, -6.0, ..., +6.0, +7.0}

The fixed point case now mimics 4 bit data that is converted to 8 bit
data by shifting left 4 bits (equivalent to multiplying by 16).  This
results in input values in the set:

{-240, -224, ..., +224, +240}

Note that the simulated input data are still uniformly distributed
rather than normally distributed, so they still differ statistically
from actual input data which are normally distributed.