Reduce dependence on std::complex<T> type
This redefines Complex and ComplexInput to be structures with two
elements: real and imag.  This makes these structures compatible with C,
whereas before they were C++ specific since they were defined using
std::complex<T>.  Being compatible with C makes the code one step closer
to a general purpose library that can be used with C or C++.

std::complex<T> is still used in omp_xengine.cc, but this file is not
intended to be part of the general purpose library.