Support sizing parameters on make command line
The sizing parameters can now be specified on the make command line if
the defaults in xgpu_info.h are not suitable for a given application.
The options that can be given on the make command line are shown here
with there default values.  Note that NTIME_PIPE must be a multiple of 4
and NTIME must be a multiple of NTIME_PIPE.  The preprocessor will error
out if those two conditions are not met.

NPOL=2
NSTATION=256
NFREQUENCY=10
NTIME=1000
NTIME_PIPE=100

For example, to compile with NSTATION set to 128 and all other
parameters at their default values:

make NSTATION=128