Make internal context's streams field an array
The internal context's streams field had been a pointer that was
initialized in xgpuInit by a call to malloc to allocate memory for two
CUDA stream structures.  This changes the streams field to be a
preallocated array of 2 and removes the malloc.  Interestingly, the
memory allocated by that malloc call was never freed, so this fixes a
memory leak as well!