Support multiple CUDA devices
This change allows xGPU to be used with multiple CUDA devices.  The call
to xgpuInit now requires a device number.  The device number is stored
in the internal context.  Every call that interacts with the CUDA
device, now calls setCudaDevice with the device number from the internal
context.  This supports multiple processes each with their own CUDA
device as well as a single process with multiple CUDA devices.  Note
that no thread safety is performed in xGPU.  If the client process has
multiple threads interacting with the xGPU library, then it is the
client's responsiblity to ensure that multiple threads do not overlap
calls to xGPU functions.