Add input/output buffer offsets to xgpuCudaXengine
Previously, xgpuCudaXengine assumed that the start of the input buffer
was the start of the input data to be correlated and that the
input buffer was exaclty one input vector long.  Likewise, it assumed
that the start of the output buffer was where the output matrix was to
be copied and the the output buffer was exactly one output matrix long.
Users wanting to use multiple externally allocated input/output buffers
would call xgpuSetHostInputBuffer before each call to xgpuCudaXengine.
The problem with this approach is that xgpuSetHostInputBuffer calls
cudaRegisterHost and cudaUnregisterHost, which turn out to take a
surprisingly long time.  For details, see the NVIDIA Forum post entitled
"Poor performance cudaHostUnregister":

http://forums.nvidia.com/index.php?showtopic=210296

To circumvent this problem, the XGPUContext structure has been given two
new fields: array_len and matrix_len (both of type size_t).  These new
fields must be used when registering host allocated buffers to indicate
the length of the region being registered.  When using CUDA allocated
host buffers, these two new fields are set to the length of one input
vector and one output matrix, respectively.

Additionally, xgpuCudaXengine now takes an input and output offset so
that the user can do a one-time registration of one large input buffer
and one large output buffer one, each of which contain multiple input
and output buffers.  Then the user can simply pass the offsets of the
small buffers within the large buffers to xgpuCudaXengine.