Change cuda_correlator -n option
This changes cuda_correlator's -n option to dump on the last call to
xgpuCudaXengine() rather than not dump at all.  This actually exposes
a problem in xgpuCudaXengine() when called back-to-back with doDump=0.
If the calls are too close, it is possible for the second call to start
transferring data into the buffer still in use by asynchronous kernel
calls launched by the first call.