Fix back-to-back xgpuCudaXengine(doDump=0) calls
When calling xgpuCudaXengine() back-to-back too quickly with doDump=0,
it was possible that the second call would start a host-to-device
transfer before asynchronous kernel calls launched in the first call was
done with the data being overwritten.  To see this, run:

cuda_correlator -c 1
cuda_correlator -c 2 -n

The first invocation shows the expected (small) number of discrepancies
(also small) between CPU and GPU.  The second command should show the
same number of small discrepancies.  If that test is done using xGPU
versions prior to this one, then the results for the second command will
show an excessive (and inconsistent!) number of errors.

This bug was due to a lack of stream synchronization between calls to
xgpuCudaXengine() when doDump=0.  When doDump=1, the synchronous
cudaMemcpy call performs an explicit synchronization of the streams, so
overwriting a still-active buffer will not be possible.

The fix for this involves moving the "completion events" into the
internal context, waiting for the "kernel completion #0" event
before the initial host-to-device copy, and recording the "kernel
completion #1" event after the final kernel launch.