Fix Makefile anamoly with local includes
The Makefile specified "-I." so that some local include files could be
found.  This causes problems if local files are created with the same
name as a system header file (e.g. "new") because it PREpends "." to the
search path for system header files.  The solution is to fix the
includes of local files by using '#include "local.h"' rather than
'#include <local.h>'.  Using quotes instead of angle brackets
automatically searches first in the same directory as the source file,
so "-I." can be removed from Makefile.