Change output format of cuda_correlator
Cleans up formatting of the line showing the total and average elapsed
time.  If compiled with RUNTIME_STATS defined, prints an extra " final"
at the end of the line for the final call to xgpuCudaXengine since that
call always uses SYNCOP_DUMP and the earlier calls (if any) use the
user-specified syncOp (if any).  This facilitates special treatment of
that line by any post analysis software.