Show max bandwidth at end of cuda_correlator run
In addition to the total elapsed time and per call elapsed time,
cuda_correlator now displays the theoretical maximum bandwidth that
could be sustained.  The theoretical maximum bandwidth is calculated as:

  time_per_channel = time_per_call / (NTIME * NFREQUENCY)

  max_bw = 1 / time_per_channel
         = (NTIME * NFREQUENCY) / time_per_call

The reported bandwidth is for a single X engine only.  How the X engine
bandwidth relates to the overall bandwidth depends on what fraction of
the overall band is processed by a single X engine.