Only compare CPU/GPU if dumping GPU exactly once
Requesting multiple dumps helps smooth out the noise for timing tests,
but is unnecessary for comparing CPU and GPU output.  Comparing CPU and
GPU output is useful for functionality testing, but is not relevant for
timing tests.  Only doing the comparison when dumping the GPU exactly
one time allows both tests to be performed without hindering each other.